Return Term::GentooFunctions, Log::Dispatch::Screen::Gentoo:
Term::GentooFunctions finally saw a new release which does not depend on
Term::ANSIScreen and can now be easily installed.

This removes all the stuff we had in place to make it work. I've also
removed the cpanfile snapshot file. I want to recreate it but carton
doesn't play nice with the proxy so fails to install anything.