Don't replace symlinks with actual copies of files when creating parcel
The Archive::Tar::Wrapper, which we used to create parcels,
can't correctly operate with symlinks and replaces them with actual
copies of files. This behavior breaks the way how system operate with
native libraries. System uses symlinks to point to the current version
of library. So, we need symlinks.

I replaced Archive::Tar::Wrapper with Archive::Tar to fix that.