Rename all backends to lowercase:
This is a small step towards a cleaner configuration.

By lowercasing all backends, we can match each backend to a URI:

* file://... -> Pakket::Repository::Backend::file
* http://... -> Pakket::Repository::Backend::http
* dbi://...  -> Pakket::Repository::Backend::dbi

The other two alternatives are to either hard-code in a main module
all the available backends and the URIs that match them, or cycle
through every available backend with each URI until the backend
matches it (hopefully correctly).