Do not instantiate the parcel repo in the Bundler at build:
We used to do this because of a fear that we would be
chdir'ed into another directory when we want it to be used,
but that's not going to happen.

At the moment, this also breaks a small test so it's a good reason
to remove it.

If anything breaks, revert this and raise an issue.