Installer: Check existence of requested packages in parcel repo
Check existence of requested packages in parcel repo before starting
the actual installation.
If one of packages doesn't exist - immediately exit with error.
Currently we have the situation if we install 2000 packages and
last one of them doesn't exist, we download and install 1999 packages
and then throw it all away, and will repeat all same work on next call,
which is bad.