Fix option --force in installer
This option didn't work because package was in the list of installed
packages and was skipped out on later stages. To make it work properly
I just delete forced packages from the list of installed and then call
the regular installation procedure.

The bug was since introducing the list of installed packages
904c75631ec1d5d776cc9b4d4839db18f81a13b5